feat!: use binary SSTable format
